Fixed POOL_waitAll bug

+++ radix_engine.h
@@ -911,7 +911,7 @@ RMF_bitpackBuildTable
 RMF_structuredBuildTable
 #endif
 (FL2_matchTable* const tbl,
-    unsigned const job,
+	size_t const job,
     unsigned const multi_thread,
     FL2_dataBlock const block,
     FL2_progressFn progress, void* opaque, U32 weight, size_t init_done)
